Fixes a possible java.lang.IllegalStateException
Based on links at final of description, a error can happen if there is
conflicted itemId in dataset. It may be caused if the adapter implements
a header or footer, when getItemId() returns 0.

Now, when is requested an id for a invalid position, a
RecyclerView.NO_ID is returned.

Links:

https://github.com/h6ah4i/android-advancedrecyclerview/issues/164
https://github.com/cymcsg/UltimateRecyclerView/issues/147